# docs/swagger.yaml
basePath: /
definitions:
...
  model.PushWaIncoming:
    properties:
      chat_id:
        type: string
      deviceNumber:
        type: string
      file_url:
        description: Versi 1
        type: string
      from:
        type: string
      message:
        type: string
      mime_type:
        description: Versi 2
        type: string
      mimetype:
        description: Versi 1
        type: string
      participant:
        type: string
      pushname:
        type: string
      remote_jid:
        type: string
      url:
        description: Versi 2 (Sering dipakai)
        type: string
    type: object
  model.Response:
    properties:
      response:
        type: string
    type: object
host: kawai-be.vercel.app
info:
  contact:
    name: Dwi Puspa Firdaus
  description: Dokumentasi API Backend untuk Kawai Assistant (Bot WhatsApp & Dashboard).
  title: Kawai Assistant API
  version: 1.0.0
paths:
...
  /webhook/nomor/{nomor}:
    post:
      consumes:
      - application/json
      description: Endpoint untuk menerima pesan masuk dari WhatsApp melalui webhook.
      parameters:
      - description: Nomor WhatsApp bot / penerima
        in: path
        name: nomor
        required: true
        type: string
      - description: Payload pesan WhatsApp dari PushWa
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.PushWaIncoming'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: Webhook WhatsApp (PushWa)
      tags:
      - Webhook
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"